\name{flexrsurv -internal.Rd}
\alias{flexrsurv -internal.Rd}
\alias{.SplineBasis-class}
\alias{getDegree,.SplineBasis-method}
\alias{getKnots,.SplineBasis-method}
\alias{getLog,.SplineBasis-method}
\alias{getNBases,.SplineBasis-method}
\alias{getOrder,.SplineBasis-method}
\alias{*-methods}
\alias{*,MSplineBasis,numeric-method}
\alias{*,numeric,MSplineBasis-method}
\alias{*,numeric,SplineBasis-method}
\alias{*,numeric,TPSplineBasis-method}
\alias{*,SplineBasis,numeric-method}
\alias{*,TPSplineBasis,numeric-method}
\alias{+-methods}
\alias{+,MSplineBasis,MSplineBasis-method}
\alias{+,MSplineBasis,numeric-method}
\alias{+,numeric,MSplineBasis-method}
\alias{+,numeric,SplineBasis-method}
\alias{+,numeric,TPSplineBasis-method}
\alias{+,SplineBasis,numeric-method}
\alias{+,SplineBasis,SplineBasis-method}
\alias{+,TPSplineBasis,numeric-method}
\alias{+,TPSplineBasis,TPSplineBasis-method}
\alias{--methods}
\alias{-,MSplineBasis,MSplineBasis-method}
\alias{-,MSplineBasis,numeric-method}
\alias{-,numeric,MSplineBasis-method}
\alias{-,numeric,SplineBasis-method}
\alias{-,numeric,TPSplineBasis-method}
\alias{-,SplineBasis,numeric-method}
\alias{-,SplineBasis,SplineBasis-method}
\alias{-,TPSplineBasis,numeric-method}
\alias{*,LEMSplineBasis,numeric-method}
\alias{*,numeric,LEMSplineBasis-method}
\alias{+,LEMSplineBasis,LEMSplineBasis-method}
\alias{+,LEMSplineBasis,numeric-method}
\alias{+,numeric,LEMSplineBasis-method}
\alias{-,LEMSplineBasis,LEMSplineBasis-method}
\alias{-,LEMSplineBasis,numeric-method}
\alias{-,numeric,LEMSplineBasis-method}
\alias{\%*\%,BSplineBasis,matrix-method}
\alias{\%*\%,LEBSplineBasis,matrix-method}
\alias{\%*\%,SplineBasis,matrix-method}
\alias{\%*\%,matrix,BSplineBasis-method}
\alias{\%*\%,matrix,LEBSplineBasis-method}
\alias{\%*\%,matrix,SplineBasis-method}
\alias{*,BSplineBasis,numeric-method}
\alias{*,LEBSplineBasis,numeric-method}
\alias{*,numeric,BSplineBasis-method}
\alias{*,numeric,LEBSplineBasis-method}
\alias{+,BSplineBasis,BSplineBasis-method}
\alias{+,BSplineBasis,numeric-method}
\alias{+,LEBSplineBasis,LEBSplineBasis-method}
\alias{+,LEBSplineBasis,numeric-method}
\alias{+,numeric,BSplineBasis-method}
\alias{+,numeric,LEBSplineBasis-method}
\alias{-,BSplineBasis,BSplineBasis-method}
\alias{-,BSplineBasis,numeric-method}
\alias{-,LEBSplineBasis,LEBSplineBasis-method}
\alias{-,LEBSplineBasis,numeric-method}
\alias{-,numeric,BSplineBasis-method}
\alias{-,numeric,LEBSplineBasis-method}
\alias{dim,BSplineBasis-method}
\alias{dim,EBSplineBasis-method}


\title{
Internal Objects and functions for the \pkg{flexrsurv } package
}
\description{
  These are not to be called by the user.
}

\author{ Michel Grzebyk \email{michel.grzebyk@inrs.fr}}

\keyword{misc}
