\name{flexrsurv-internal.Rd}
\alias{flexrsurv-internal.Rd}
\alias{flexrsurv.ll.fromto.brass.wce.fit}
\alias{flexrsurv.ll.fromto.brass0.wceadd.fitCoptim}
\alias{flexrsurv.ll.fromto.brass0.wce.fitCoptim}
\alias{flexrsurv.ll.fromto.brass0.periode.wce.fitCoptim}
\alias{weighted_cummulative_exposure}
\alias{MSplineBasis-class}
\alias{MSplineBasis}
\alias{gradientwce}
\alias{predictwce}

\alias{ndf.flexrsurvclt}


\alias{.SplineBasis-class}
\alias{getDegree,.SplineBasis-method}
\alias{getKnots,.SplineBasis-method}
\alias{getLog,.SplineBasis-method}
\alias{getNBases,.SplineBasis-method}
\alias{getOrder,.SplineBasis-method}
\alias{*-methods}
\alias{*,MSplineBasis,numeric-method}
\alias{*,numeric,MSplineBasis-method}
\alias{*,numeric,SplineBasis-method}
\alias{*,numeric,TPSplineBasis-method}
\alias{*,SplineBasis,numeric-method}
\alias{*,TPSplineBasis,numeric-method}
\alias{+-methods}
\alias{+,MSplineBasis,MSplineBasis-method}
\alias{+,MSplineBasis,numeric-method}
\alias{+,numeric,MSplineBasis-method}
\alias{+,numeric,SplineBasis-method}
\alias{+,numeric,TPSplineBasis-method}
\alias{+,SplineBasis,numeric-method}
\alias{+,SplineBasis,SplineBasis-method}
\alias{+,TPSplineBasis,numeric-method}
\alias{+,TPSplineBasis,TPSplineBasis-method}
\alias{--methods}
\alias{-,MSplineBasis,MSplineBasis-method}
\alias{-,MSplineBasis,numeric-method}
\alias{-,numeric,MSplineBasis-method}
\alias{-,numeric,SplineBasis-method}
\alias{-,numeric,TPSplineBasis-method}
\alias{-,SplineBasis,numeric-method}
\alias{-,SplineBasis,SplineBasis-method}
\alias{-,TPSplineBasis,numeric-method}
\alias{*,LEMSplineBasis,numeric-method}
\alias{*,numeric,LEMSplineBasis-method}
\alias{+,LEMSplineBasis,LEMSplineBasis-method}
\alias{+,LEMSplineBasis,numeric-method}
\alias{+,numeric,LEMSplineBasis-method}
\alias{-,LEMSplineBasis,LEMSplineBasis-method}
\alias{-,LEMSplineBasis,numeric-method}
\alias{-,numeric,LEMSplineBasis-method}
\alias{\%*\%,BSplineBasis,matrix-method}
\alias{\%*\%,LEBSplineBasis,matrix-method}
\alias{\%*\%,SplineBasis,matrix-method}
\alias{\%*\%,matrix,BSplineBasis-method}
\alias{\%*\%,matrix,LEBSplineBasis-method}
\alias{\%*\%,matrix,SplineBasis-method}
\alias{*,BSplineBasis,numeric-method}
\alias{*,LEBSplineBasis,numeric-method}
\alias{*,numeric,BSplineBasis-method}
\alias{*,numeric,LEBSplineBasis-method}

\alias{*,numeric,BDSplineBasis-method}
\alias{*,BDSplineBasis,numeric-method}
\alias{*,numeric,MDSplineBasis-method}
\alias{*,MDSplineBasis,numeric-method}

\alias{+,BSplineBasis,BSplineBasis-method}
\alias{+,BSplineBasis,numeric-method}
\alias{+,LEBSplineBasis,LEBSplineBasis-method}
\alias{+,LEBSplineBasis,numeric-method}
\alias{+,numeric,BSplineBasis-method}
\alias{+,numeric,LEBSplineBasis-method}

\alias{+,BDSplineBasis,BDSplineBasis-method}
\alias{+,numeric,BDSplineBasis-method}
\alias{+,BDSplineBasis,numeric-method}
\alias{+,MDSplineBasis,MDSplineBasis-method}
\alias{+,numeric,MDSplineBasis-method}
\alias{+,MDSplineBasis,numeric-method}

\alias{-,BSplineBasis,BSplineBasis-method}
\alias{-,BSplineBasis,numeric-method}
\alias{-,LEBSplineBasis,LEBSplineBasis-method}
\alias{-,LEBSplineBasis,numeric-method}
\alias{-,numeric,BSplineBasis-method}
\alias{-,numeric,LEBSplineBasis-method}

\alias{-,BDSplineBasis,BDSplineBasis-method}
\alias{-,numeric,BDSplineBasis-method}
\alias{-,BDSplineBasis,numeric-method}
\alias{-,MDSplineBasis,MDSplineBasis-method}
\alias{-,numeric,MDSplineBasis-method}
\alias{-,MDSplineBasis,numeric-method}

\alias{dim,BSplineBasis-method}
\alias{dim,EBSplineBasis-method}

\alias{gradientwce, TPSplineBasis,numeric,numeric,numeric-method}
\alias{gradientwce, MSplineBasis,numeric,numeric,numeric-method}
\alias{gradientwce, MDSplineBasis,numeric,numeric,numeric-method}
\alias{gradientwce, EMSplineBasis,numeric,numeric,numeric-method}


\alias{predictwce, TPSplineBasis,numeric,numeric,numeric-method}
\alias{predictwce, MSplineBasis ,numeric,numeric,numeric-method}
\alias{predictwce, MDSplineBasis,numeric,numeric,numeric-method}
\alias{predictwce, EMSplineBasis,numeric,numeric,numeric-method}
\alias{predictwce, BSplineBasis ,numeric,numeric,numeric-method}
\alias{predictwce, BDSplineBasis,numeric,numeric,numeric-method}


\alias{predictSpline,BDSplineBasis,numeric-method}
\alias{predictSpline,BSplineBasis,numeric-method}
\alias{predictSpline,EBSplineBasis,numeric-method}
\alias{predictSpline,LEBSplineBasis,numeric-method}
\alias{predictSpline,R2BSplineBasis,numeric-method}
\alias{predictSpline,R2bBSplineBasis,numeric-method}
\alias{predictSpline,TPSplineBasis,numeric-method}


\title{
Internal Objects and functions for the \pkg{flexrsurv } package
}
\description{
  These are not to be called by the user.
}

\author{ Michel Grzebyk \email{michel.grzebyk@inrs.fr}}

\keyword{misc}
\keyword{internal}
